+++ source/SaveLoad.h
@@ -3,7 +3,6 @@
 #include "Entity.h"
 #include "Globals.h"
 #include "Player.h"
-#include <3ds.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -17,4 +16,4 @@ extern bool saveFileCopy(char *target, char *source);
 extern bool saveWorld(char *filename, EntityManager *eManager, WorldData *worldData, PlayerData *players, int playerCount);
 extern bool loadWorld(char *filename, EntityManager *eManager, WorldData *worldData, PlayerData *players, int playerCount);
 
-extern s8 checkFileNameForErrors(char *filename);
+extern int checkFileNameForErrors(char *filename);
