+++ source/network/Synchronizer.h
@@ -1,6 +1,7 @@
 #pragma once
 
-#include <3ds.h>
+#include <stdbool.h>
+#include <stddef.h>
 
 // 2-3 seem be optimal (at least for 2 players)
 #define SYNCHRONIZER_TICKS_PER_TURN 2
@@ -8,7 +9,7 @@
 extern void synchronizerInit(int seed, int initPlayerCount, int initPlayerLocalID);
 
 extern void synchronizerSendUID();
-extern void synchronizerSetPlayerUID(int playerID, u32 uid);
+extern void synchronizerSetPlayerUID(int playerID, int uid);
 extern void synchronizerSendIfReady();
 extern void synchronizerSetPlayerReady(int playerID);
 extern bool synchronizerAllReady();
@@ -20,10 +21,10 @@ extern void synchronizerTick(void (*gtick)(void));
 
 extern void synchronizerReset();
 
-extern void synchronizerOnInputPacket(u8 playerID, u32 turnNumber, void *data, size_t dataSize);
+extern void synchronizerOnInputPacket(int playerID, int turnNumber, void *data, size_t dataSize);
 
 // values used ingame
-extern u32 syncTickCount;
+extern int syncTickCount;
 
 // helpers for random numbers
 extern double gaussrand(bool reset);
